Refactor: Extract MotionSequenceHandler and BlinkConditionHandler into separate modules
- Created motion_sequence_handler.py for motion sequence processing logic
- Created blink_condition_handler.py for blink state handling
- Updated all imports across the codebase to use new module structure
- Cleaned up vasa_model.py by removing ~500 lines of code
- Better separation of concerns and modularity
